Extend examples of subnet parameter
The parameter subnets isn't well named in my opinion. I was expecting it to require the name of a AWS subnet until I checked the configuration of the already existing default-InstanceGroup. Only then I noticed that subnets was meant to be the availability zone or a list of AZs. This change is supposed to make that more obvious.